Improve zarr specs for data consistency

+++ tests/test_csr_to_csc.py
@@ -36,8 +36,8 @@ def _make_sparse_adata(
     feature_uids: list[str],
     rng: np.random.Generator,
 ) -> ad.AnnData:
-    X = sp.random(n_obs, n_vars, density=0.3, format="csr", dtype=np.float32, random_state=rng)
-    X.data[:] = rng.integers(1, 100, size=X.nnz).astype(np.float32)
+    X = sp.random(n_obs, n_vars, density=0.3, format="csr", dtype=np.uint32, random_state=rng)
+    X.data[:] = rng.integers(1, 100, size=X.nnz).astype(np.uint32)
     var = pl.DataFrame({"global_feature_uid": feature_uids}).to_pandas()
     return ad.AnnData(X=X, var=var)
